parser_for_module_1-main/requirements.txt
aspose_words==25.2.0
beautifulsoup4==4.13.3
pdf2image==1.17.0
Pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python_docx==1.1.2
Requests==2.32.3
tabula_py==2.10.0
lxml
